refactor: extract jobs-pane.tsx into focused modules (2489→1599 lines)
Split the oversized jobs-pane.tsx into smaller, focused files:
- jobs-utils.tsx: shared utilities, status helpers, and form components
- jobs-add-dialog.tsx: AddJobDialog and AddJobFlow
- jobs-charts.tsx: DailyRunsChart, JobAvgDuration, RunHistoryGrid

Pure structural refactoring — no behavior changes.
